Refactor `create_model`
