+++ pumaguard/classify.py
@@ -6,7 +6,6 @@ This script classifies images.
 
 import argparse
 import datetime
-import os
 import numpy as np
 import tensorflow as tf  # type: ignore
 import keras  # type: ignore
@@ -14,11 +13,9 @@ import keras  # type: ignore
 from pumaguard.presets import Presets
 from pumaguard.traininghistory import TrainingHistory
 from pumaguard.utils import (
+    create_model,
     initialize_tensorflow,
 )
-from pumaguard.models.pretrained import pre_trained_model
-from pumaguard.models.light import light_model
-from pumaguard.models.light_2 import light_model_2
 
 
 def get_duration(start_time: datetime.datetime,
@@ -35,64 +32,6 @@ def get_duration(start_time: datetime.datetime,
     """
     duration = end_time - start_time
     return duration / datetime.timedelta(microseconds=1) / 1e6
-
-
-def create_model(presets: Presets,
-                 distribution_strategy: tf.distribute.Strategy):
-    """
-    Create the model.
-    """
-    with distribution_strategy.scope():
-        model_file_exists = os.path.isfile(presets.model_file)
-        if presets.load_model_from_file and model_file_exists:
-            os.stat(presets.model_file)
-            print(f'Loading model from file {presets.model_file}')
-            model = keras.models.load_model(presets.model_file)
-            print('Loaded model from file')
-        else:
-            print('Creating new model')
-            if presets.model_version == "pre-trained":
-                print('Creating new Xception model')
-                model = pre_trained_model(presets)
-                print('Building pre-trained model')
-                model.build(input_shape=(None, *presets.image_dimensions, 3))
-                print('Compiling pre-trained model')
-                model.compile(
-                    optimizer=keras.optimizers.Adam(learning_rate=1e-4),
-                    loss='binary_crossentropy',
-                    metrics=['accuracy'],
-                )
-            elif presets.model_version == "light":
-                print('Creating new light model')
-                model = light_model(presets)
-                print('Building light model')
-                model.build(input_shape=(None, *presets.image_dimensions, 1))
-                print('Compiling light model')
-                model.compile(
-                    optimizer=keras.optimizers.Adam(
-                        learning_rate=presets.alpha),
-                    loss=keras.losses.BinaryCrossentropy(from_logits=True),
-                    metrics=[keras.metrics.BinaryAccuracy(name="accuracy")],
-                )
-            elif presets.model_version == 'light-2':
-                print('Creating new light-2 model')
-                model = light_model_2(presets)
-                print('Building light-2 model')
-                model.build(input_shape=(None, *presets.image_dimensions, 1))
-                print('Compiling light model')
-                model.compile(
-                    optimizer=keras.optimizers.Adam(learning_rate=1e-4),
-                    loss='binary_crossentropy',
-                    metrics=['accuracy'],
-                )
-            else:
-                raise ValueError(
-                    f'unknown model version {presets.model_version}')
-
-            print(f'Number of layers in the model: {len(model.layers)}')
-            model.summary()
-
-    return model
 
 
 def classify_images(presets: Presets, model: keras.Model, image_path: str):
